feat: finish authorization

+++ backend/routes/posts.js
@@ -29,7 +29,8 @@ router.post('',
     const post = new Post({
         title: req.body.title,
         content: req.body.content,
-        imagePath: url + '/images/' + req.file.filename
+        imagePath: url + '/images/' + req.file.filename,
+        creator: req.userData.userId
     });
 
     post.save()
@@ -62,7 +63,6 @@ router.get('', (req, res, next) => {
     
     postQuery
         .then(documents => {
-            
             fetchedPosts = documents;
             return Post.countDocuments();
         })
@@ -99,9 +99,13 @@ router.get('/:id', (req, res, next) => {
 router.delete('/:id', 
     checkAuth,
     (req, res, next) => {
-    Post.deleteOne({ _id: req.params.id })
+    Post.deleteOne({ _id: req.params.id, creator: req.userData.userId })
         .then(result => {
-            res.status(200).json({ message: 'Post deleted!' });
+            if (result.deletedCount > 0) {
+                res.status(200).json({ message: 'Post deleted!' });
+            } else {
+                res.status(401).json({ message: 'Not authorized!' });
+            }
         })
         .catch(error => {
             res.status(500).json({
@@ -128,12 +132,17 @@ router.put('/:id',
         _id: req.body.id,
         title: req.body.title,
         content: req.body.content,
-        imagePath
+        imagePath,
+        creator: req.userData.userId
     });
 
-    Post.updateOne({ _id: req.params.id }, post)
+    Post.updateOne({ _id: req.params.id, creator: req.userData.userId }, post)
         .then(result => {
-            res.status(200).json({ message: 'Update successful!' });
+            if (result.modifiedCount > 0) {
+                res.status(200).json({ message: 'Update successful!' });
+            } else {
+                res.status(401).json({ message: 'Not authorized!' });
+            }
         })
         .catch(error => {
             res.status(500).json({
